combine test cases
