--- tests/integration/telemetry/stats/prometheus/extsionproviders/accesslogs_test.go
@@ -1,62 +1,0 @@
-//go:build integ
-// +build integ
-
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package extsionproviders
-
-import (
-	"testing"
-
-	"istio.io/istio/pkg/test/framework"
-	"istio.io/istio/pkg/test/framework/components/istio"
-	"istio.io/istio/pkg/test/framework/label"
-	"istio.io/istio/pkg/test/framework/resource"
-	common "istio.io/istio/tests/integration/telemetry/stats/prometheus"
-)
-
-func TestMain(m *testing.M) {
-	framework.NewSuite(m).
-		Label(label.CustomSetup).
-		Label(label.IPv4). // https://github.com/istio/istio/issues/35915
-		Setup(istio.Setup(common.GetIstioInstance(), setupConfig)).
-		Setup(common.TestSetup).
-		Run()
-}
-
-func setupConfig(_ resource.Context, cfg *istio.Config) {
-	if cfg == nil {
-		return
-	}
-	cfg.ControlPlaneValues = `
-meshConfig:
-  accessLogFile: ""
-  extensionProviders:
-  - name: file-log
-    envoyFileAccessLog:
-      path: /dev/stdout
-  defaultProviders:
-    accessLogging:
-    - file-log
-`
-}
-
-func TestAccessLogsDefaultProvider(t *testing.T) {
-	framework.NewTest(t).
-		Features("observability.telemetry.logging.defaultprovider").
-		Run(func(t framework.TestContext) {
-			common.RunAccessLogsTests(t, true)
-		})
-}
